Create CMakeLists in subdirectory, restructure tests/
