+++ test/doctest-main.cpp
@@ -1,0 +1,10 @@
+//#define CATCH_CONFIG_MAIN
+//#include "doctest.h"
+#define DOCTEST_CONFIG_IMPLEMENT
+#include "doctest/doctest.h"
+
+#include <stdlib.h>
+
+int main(int argc, char** argv) {
+  return 0;
+}
