+++ test/encoder/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_executable(test-ogg-encoder ogg-encoder-tests.cpp)
+
+target_compile_features(test-ogg-encoder PRIVATE cxx_std_17)
+target_link_libraries(test-ogg-encoder PRIVATE image_library doctest)
+#target_link_libraries(test-ogg-encoder PRIVATE image_library doctest::doctest)
+
+#target_link_libraries(test-ogg-encoder PRIVATE image_library ${DOCTEST_INCLUDE_DIR})
+target_include_directories(test-ogg-encoder PUBLIC ${DOCTEST_INCLUDE_DIR} ) 
+
+add_test(NAME testencodertest COMMAND test-ogg-encoder)
