--- test/encoder/main/ogg-encoder-tests.cpp
@@ -1,12 +1,0 @@
-#include "catch.hpp"
-
-#include <string>
-
-bool embedInImage(std::string imgPath, std::string oggPath) {
-  return embedInImage(imgPath, oggPath);
-}
-
-
-TEST_CASE( "OGG file is embedded in image" ,  "[encoder]") {
-  // Fill in with files of various formats.
-}
